# services/document_resolution_api/tests/test_public_api/api/_snapshots/openapi.resolve.v1.json
{
  "components": {
    "schemas": {
...
      "DefinedTermBlock": {
        "properties": {
          "candidate_scores": {
            "description": "Ranked candidate meanings considered for this occurrence.",
            "items": {
              "$ref": "#/components/schemas/DefinedTermCandidate"
            },
            "title": "Candidate Scores",
            "type": "array"
          },
          "chosen_definition_span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TextSpan"
              },
              {
                "type": "null"
              }
            ],
            "description": "Definition span associated with the selected meaning when available."
          },
          "chosen_meaning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DefinedTermMeaningRef"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved meaning record when available."
          },
          "chosen_meaning_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Selected meaning identifier, or null when unresolved.",
            "title": "Chosen Meaning Id"
          },
          "normalized_key": {
            "description": "Canonical normalized key for the occurrence.",
            "title": "Normalized Key",
            "type": "string"
          },
          "occurrence_span": {
            "$ref": "#/components/schemas/TextSpan",
            "description": "Span of the later defined-term occurrence."
          },
          "resolution_method": {
            "description": "Resolution path used for the final decision.",
            "enum": [
              "tier1",
              "tier2_blend",
              "unresolved"
            ],
            "title": "Resolution Method",
            "type": "string"
          },
          "resolved": {
            "description": "True when a meaning was selected.",
            "title": "Resolved",
            "type": "boolean"
          },
          "term": {
            "description": "Surface text of the occurrence.",
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "occurrence_span",
          "term",
          "normalized_key",
          "resolution_method",
          "resolved"
        ],
        "title": "DefinedTermBlock",
        "type": "object"
      },
...
      "Definition": {
        "properties": {
          "confidence": {
            "description": "0\u20131 confidence score for this definition.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "End offset (exclusive).",
            "title": "End"
          },
          "source": {
            "$ref": "#/components/schemas/DefinitionSource",
            "description": "Origin of the definition: extracted|glossary."
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Start offset of definition text if found in source text.",
            "title": "Start"
          },
          "text": {
            "description": "Expanded form of the acronym.",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "confidence",
          "source"
        ],
        "title": "Definition",
        "type": "object"
      },
      "DefinitionSource": {
        "enum": [
          "extracted",
          "glossary"
        ],
        "title": "DefinitionSource",
        "type": "string"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine-readable enum."
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured diagnostics (limit, actual, etc.).",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable error.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "GlossaryBlock": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/GlossaryMatch"
            },
            "title": "Matches",
            "type": "array"
          }
        },
        "title": "GlossaryBlock",
        "type": "object"
      },
...
      "OrchestrationMeta": {
        "properties": {
          "completed": {
            "items": {
              "type": "string"
            },
            "title": "Completed",
            "type": "array"
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "title": "Failed",
            "type": "array"
          },
          "requested": {
            "items": {
              "type": "string"
            },
            "title": "Requested",
            "type": "array"
          }
        },
        "required": [
          "requested",
          "completed",
          "failed"
        ],
        "title": "OrchestrationMeta",
        "type": "object"
      },
      "PipelineError": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/PipelineErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "pipeline": {
            "title": "Pipeline",
            "type": "string"
          }
        },
        "required": [
          "pipeline",
          "code",
          "message"
        ],
        "title": "PipelineError",
        "type": "object"
      },
...
      "ResolutionMode": {
        "description": "Deterministic resolution policy for selecting a final acronym meaning.\n\nSTRICT:\n    Select only when the resolver can do so conservatively. Avoid permissive\n    fallback behaviour and prefer leaving the acronym unresolved when\n    multiple glossary candidates remain.\n\nDOMAIN_PRIORITY:\n    Use the standard deterministic policy. Prefer the strongest available\n    candidate according to the resolver's normal ordering and fallback\n    rules. This is the default balanced mode.\n\nFALLBACK_GENERAL:\n    Prefer returning a usable result whenever possible. If no stronger\n    candidate is available, prefer a general-domain meaning before falling\n    back to the first deterministically ordered candidate.",
        "enum": [
          "strict",
          "domain_priority",
          "fallback_general"
        ],
        "title": "ResolutionMode",
        "type": "string"
      },
      "ResolveCandidate": {
        "properties": {
          "definition": {
            "description": "Candidate resolved meaning.",
            "title": "Definition",
            "type": "string"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Meaning/domain tag if known, else null.",
            "title": "Domain"
          },
          "provenance": {
            "$ref": "#/components/schemas/CandidateProvenance",
            "description": "Where the candidate came from."
          },
          "score": {
            "description": "Deterministic ranking score. In the MVP this is primarily used for candidate ordering, not calibrated confidence.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional stable source reference, e.g. text span or glossary meaning id.",
            "title": "Source Ref"
          }
        },
        "required": [
          "definition",
          "score",
          "provenance"
        ],
        "title": "ResolveCandidate",
        "type": "object"
      },
      "ResolveMeta": {
        "properties": {
          "input_chars": {
            "title": "Input Chars",
            "type": "integer"
          },
          "model_version": {
            "description": "plainera-core version used for processing.",
            "title": "Model Version",
            "type": "string"
          },
          "processing_ms": {
            "title": "Processing Ms",
            "type": "integer"
          },
          "resolution_mode": {
            "$ref": "#/components/schemas/ResolutionMode",
            "description": "Resolution policy applied during sense selection."
          }
        },
        "required": [
          "processing_ms",
          "model_version",
          "input_chars",
          "resolution_mode"
        ],
        "title": "ResolveMeta",
        "type": "object"
      },
      "ResolveOptions": {
        "properties": {
          "include_glossary_enrichment": {
            "default": true,
            "description": "If true, attempt curated joins from repository (read-only).",
            "title": "Include Glossary Enrichment",
            "type": "boolean"
          },
          "locale": {
            "default": "en-GB",
            "description": "Locale hint for heuristics. ISO BCP-47 tag. Supported: en-GB, en-US.",
            "title": "Locale",
            "type": "string"
          },
          "max_definitions_per_acronym": {
            "default": 5,
            "description": "Maximum candidate definitions/meanings to return per acronym.",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Definitions Per Acronym",
            "type": "integer"
          },
          "min_confidence": {
            "default": 0.0,
            "description": "Drop candidates below this confidence threshold.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Confidence",
            "type": "number"
          },
          "return_occurrences": {
            "default": true,
            "description": "If true, include all {start,end} positions.",
            "title": "Return Occurrences",
            "type": "boolean"
          },
          "window_chars": {
            "default": 120,
            "description": "Context window size used in responses.",
            "minimum": 0.0,
            "title": "Window Chars",
            "type": "integer"
          }
        },
        "title": "ResolveOptions",
        "type": "object"
      },
      "ResolveRequest": {
        "examples": [
          {
            "options": {
              "include_glossary_enrichment": true,
              "locale": "en-GB",
              "max_definitions_per_acronym": 5,
              "min_confidence": 0.0,
              "return_occurrences": true,
              "window_chars": 120
            },
            "resolution_mode": "domain_priority",
            "targets": [
              "acronyms",
              "defined_terms",
              "structural_references"
            ],
            "text": "The Metropolitan Police Service (MPS) operates in London."
          },
          {
            "text": ""
          }
        ],
        "properties": {
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolveOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolution_mode": {
            "$ref": "#/components/schemas/ResolutionMode",
            "default": "domain_priority",
            "description": "Controls deterministic acronym sense selection policy."
          },
          "targets": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ResolveTarget"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional explicit pipeline targets to run. If omitted, all supported targets are selected.",
            "title": "Targets"
          },
          "text": {
            "description": "Raw document content. Max length 100,000 characters.",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ResolveRequest",
        "type": "object"
      },
      "ResolveResponse": {
        "examples": [
          {
            "acronyms": [
              {
                "acronym": "GP",
                "candidates": [
                  {
                    "definition": "General Practitioner",
                    "provenance": "document",
                    "score": 1.0,
                    "source_ref": "text_span:4-24"
                  },
                  {
                    "definition": "General Partner",
                    "domain": "finance",
                    "provenance": "glossary",
                    "score": 0.0,
                    "source_ref": "meaning:42"
                  }
                ],
                "conflict": true,
                "conflict_count": 2,
                "definitions": [
                  {
                    "confidence": 0.99,
                    "end": 24,
                    "source": "extracted",
                    "start": 4,
                    "text": "General Practitioner"
                  }
                ],
                "first_occurrence": {
                  "end": 35,
                  "start": 33
                },
                "glossary": {
                  "matches": [
                    {
                      "confidence": 0.98,
                      "definition": "General Practitioner",
                      "domain": "medical",
                      "lang": "en",
                      "source": "system"
                    },
                    {
                      "confidence": 0.82,
                      "definition": "General Partner",
                      "domain": "finance",
                      "lang": "en",
                      "source": "system"
                    }
                  ]
                },
                "occurrences": [
                  {
                    "end": 35,
                    "start": 33
                  }
                ],
                "selected": {
                  "definition": "General Practitioner",
                  "reason": "in_document_definition"
                },
                "selection": {
                  "filtered_inactive_count": 0
                }
              }
            ],
            "defined_terms": [
              {
                "candidate_scores": [
                  {
                    "components": {
                      "recency": 0.5,
                      "section_proximity": 0.5
                    },
                    "definition_span": {
                      "end": 96,
                      "start": 40
                    },
                    "meaning_id": "term|services|1",
                    "tier1_score": 1.0,
                    "total_score": 1.0
                  }
                ],
                "chosen_definition_span": {
                  "end": 96,
                  "start": 40
                },
                "chosen_meaning": {
                  "definition_span": {
                    "end": 96,
                    "start": 40
                  },
                  "definition_text": "the support and maintenance services described in Schedule 1",
                  "intro_kind": "quoted_means",
                  "intro_span": {
                    "end": 28,
                    "start": 20
                  },
                  "meaning_id": "term|services|1",
                  "normalized_key": "services",
                  "ordinal": 1,
                  "section_path": [
                    "1",
                    "Definitions"
                  ],
                  "surface": "Services"
                },
                "chosen_meaning_id": "term|services|1",
                "normalized_key": "services",
                "occurrence_span": {
                  "end": 128,
                  "start": 120
                },
                "resolution_method": "tier1",
                "resolved": true,
                "term": "Services"
              }
            ],
            "errors": [
              {
                "code": "PIPELINE_EXECUTION_FAILED",
                "message": "boom",
                "pipeline": "defined_terms"
              }
            ],
            "meta": {
              "input_chars": 680,
              "model_version": "plainera-core@1.0.0",
              "processing_ms": 12,
              "resolution_mode": "domain_priority"
            },
            "orchestration": {
              "completed": [
                "acronyms",
                "structural_references"
              ],
              "failed": [
                "defined_terms"
              ],
              "requested": [
                "acronyms",
                "defined_terms",
                "structural_references"
              ]
            },
            "structural_references": [
              {
                "canonical_key": "section|4.2",
                "canonical_label": "4.2",
                "kind": "Section",
                "label": "4.2",
                "match_strategy": "forward",
                "normalized_key": "section:4.2",
                "provenance": "document",
                "reference_span": {
                  "end": 221,
                  "start": 210
                },
                "resolved": true,
                "strength": 1.0,
                "target_span": {
                  "end": 438,
                  "start": 420
                }
              }
            ]
          }
        ],
        "properties": {
          "acronyms": {
            "items": {
              "$ref": "#/components/schemas/ResolvedAcronymBlock"
            },
            "title": "Acronyms",
            "type": "array"
          },
          "defined_terms": {
            "items": {
              "$ref": "#/components/schemas/DefinedTermBlock"
            },
            "title": "Defined Terms",
            "type": "array"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/PipelineError"
            },
            "title": "Errors",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResolveMeta"
          },
          "orchestration": {
            "$ref": "#/components/schemas/OrchestrationMeta"
          },
          "structural_references": {
            "items": {
              "$ref": "#/components/schemas/StructuralReferenceBlock"
            },
            "title": "Structural References",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "orchestration"
        ],
        "title": "ResolveResponse",
        "type": "object"
      },
      "ResolveTarget": {
        "enum": [
          "acronyms",
          "defined_terms",
          "structural_references"
        ],
        "title": "ResolveTarget",
        "type": "string"
      },
      "ResolvedAcronymBlock": {
        "properties": {
          "acronym": {
            "description": "Surface form exactly as detected (no lowercasing).",
            "title": "Acronym",
            "type": "string"
          },
          "candidates": {
            "description": "Ordered candidate meaning, best to worst, bounded by request cap.",
            "items": {
              "$ref": "#/components/schemas/ResolveCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "conflict": {
            "default": false,
            "description": "True when more than one viable candidate existed.",
            "title": "Conflict",
            "type": "boolean"
          },
          "conflict_count": {
            "default": 0,
            "description": "Number of viable candidates considered after filtering.",
            "minimum": 0.0,
            "title": "Conflict Count",
            "type": "integer"
          },
          "definitions": {
            "description": "Candidate definitions for this acronym (ranked by confidence).",
            "items": {
              "$ref": "#/components/schemas/Definition"
            },
            "title": "Definitions",
            "type": "array"
          },
          "first_occurrence": {
            "$ref": "#/components/schemas/Span",
            "description": "First occurrence offsets in Python-slice semantics (end exclusive)."
          },
          "glossary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GlossaryBlock"
              },
              {
                "type": "null"
              }
            ],
            "description": "Curated matches present when enrichment is enabled/available."
          },
          "occurrences": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Span"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "All occurrences when return_occurrences=true.",
            "title": "Occurrences"
          },
          "selected": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectedCandidate"
              },
              {
                "type": "null"
              }
            ],
            "description": "Chosen candidate when at least one viable meaning exists."
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionMeta"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional deterministic selection metadata."
          }
        },
        "required": [
          "acronym",
          "first_occurrence"
        ],
        "title": "ResolvedAcronymBlock",
        "type": "object"
      },
      "SelectedCandidate": {
        "properties": {
          "definition": {
            "description": "Chosen resolved meaning.",
            "title": "Definition",
            "type": "string"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Selected meaning/domain tag if known, else null.",
            "title": "Domain"
          },
          "reason": {
            "$ref": "#/components/schemas/SelectionReason",
            "description": "Deterministic selection reason."
          }
        },
        "required": [
          "definition",
          "reason"
        ],
        "title": "SelectedCandidate",
        "type": "object"
      },
...
      "Span": {
        "properties": {
          "end": {
            "description": "End offset (exclusive).",
            "title": "End",
            "type": "integer"
          },
          "start": {
            "description": "Start offset (inclusive).",
            "title": "Start",
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "Span",
        "type": "object"
      },
      "StructuralReferenceBlock": {
        "properties": {
          "canonical_key": {
            "description": "Canonical lookup key used for linking.",
            "title": "Canonical Key",
            "type": "string"
          },
          "canonical_label": {
            "description": "Canonicalized label used for deterministic matching.",
            "title": "Canonical Label",
            "type": "string"
          },
          "kind": {
            "description": "Structural reference kind, for example Section or Schedule.",
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "description": "Source-close detected label text.",
            "title": "Label",
            "type": "string"
          },
          "match_strategy": {
            "description": "Deterministic positional strategy used to resolve the link.",
            "enum": [
              "forward",
              "backward",
              "overlap",
              "unresolved"
            ],
            "title": "Match Strategy",
            "type": "string"
          },
          "normalized_key": {
            "description": "Normalized structural key.",
            "title": "Normalized Key",
            "type": "string"
          },
          "provenance": {
            "description": "Provenance tag for the structural reference.",
            "title": "Provenance",
            "type": "string"
          },
          "reference_span": {
            "$ref": "#/components/schemas/Span",
            "description": "Occurrence span of the structural reference."
          },
          "resolved": {
            "description": "True when the occurrence resolved to a target span.",
            "title": "Resolved",
            "type": "boolean"
          },
          "strength": {
            "description": "Deterministic link-strength score. Ordinal heuristic, not calibrated confidence.",
            "minimum": 0.0,
            "title": "Strength",
            "type": "number"
          },
          "target_span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved target heading span, or null if unresolved."
          }
        },
        "required": [
          "kind",
          "label",
          "canonical_label",
          "normalized_key",
          "canonical_key",
          "reference_span",
          "match_strategy",
          "strength",
          "provenance",
          "resolved"
        ],
        "title": "StructuralReferenceBlock",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "document_resolution API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/resolve": {
      "post": {
        "description": "Resolve acronyms, defined terms, and structural references from raw document text. The response may include occurrence-level results, deterministic resolution metadata, and structured partial-success reporting by pipeline. Offsets use Python-slice semantics (end exclusive). Supported locales: en-GB, en-US. Idempotent: does not mutate server state. Content-Encoding: gzip supported.",
        "operationId": "resolve_document_v1_resolve_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Body-Limit-Bytes": {
                "description": "Current configured body size limit.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-Input-Bytes": {
                "description": "Parsed request body size in bytes.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Limit": {
                "description": "Requests allowed in window.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Remaining": {
                "description": "Remaining in current window.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Reset": {
                "description": "Epoch seconds until window resets.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-Request-ID": {
                "description": "Echoed or generated correlation id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Resolve acronyms, defined terms, and structural references in raw text",
        "tags": [
          "Resolve"
        ]
      }
    }
  }
}